requirements.txt
streamlit==1.39.0
pymongo==4.8.0
numpy==1.26.4
pandas==2.2.2
matplotlib==3.9.2
python-docx==1.1.2
streamlit-cookies-manager==0.2.0
python-dotenv==1.0.1
easyocr==1.7.2
fastapi==0.115.2
uvicorn==0.32.0
fuzzywuzzy==0.18.0
joblib==1.4.2
faiss-cpu
sentence-transformers==3.2.1
opencv-python-headless==4.5.5.64